Convert request collector to structlog
