+++ src/monitoring_service/cli.py
@@ -1,5 +1,4 @@
 import json
-import logging
 import os
 import sys
 from typing import Any, Optional
@@ -13,6 +12,7 @@ from web3 import HTTPProvider, Web3
 from monitoring_service.constants import DEFAULT_REQUIRED_CONFIRMATIONS
 from monitoring_service.service import MonitoringService
 from raiden_contracts.contract_manager import ContractManager, contracts_precompiled_path
+from raiden_libs.logging import setup_logging
 from raiden_libs.types import Address
 
 log = structlog.get_logger(__name__)
@@ -25,34 +25,6 @@ def validate_address(_ctx: Any, _param: Any, value: Optional[str]) -> Optional[s
     if not is_checksum_address(value):
         raise click.BadParameter('not an EIP-55 checksummed address')
     return value
-
-
-def setup_logging(log_level: str) -> None:
-    logging.basicConfig(
-        level=log_level,
-        stream=sys.stdout,
-        format="%(message)s",
-    )
-
-    logging.getLogger('web3').setLevel('INFO')
-    logging.getLogger('urllib3').setLevel('INFO')
-
-    chain = [
-        structlog.stdlib.add_log_level,
-        structlog.stdlib.add_logger_name,
-        structlog.stdlib.PositionalArgumentsFormatter(),
-        structlog.processors.TimeStamper(fmt="%Y-%m-%d %H:%M:%S.%f"),
-        structlog.processors.StackInfoRenderer(),
-        structlog.processors.format_exc_info,
-        structlog.dev.ConsoleRenderer(),
-    ]
-    structlog.configure_once(
-        processors=chain,
-        context_class=dict,
-        logger_factory=structlog.stdlib.LoggerFactory(),
-        wrapper_class=structlog.stdlib.BoundLogger,
-        cache_logger_on_first_use=True,
-    )
 
 
 @click.command()
